1. Separate symbol execution from loop structure 2. Classify tests into subfolders based on loop type 3. Improve comments 4. Minor fixes
